drafts/sofieversion/haikara-agent final/requirements.txt
requests
python-dotenv
pydantic
flask
flask-cors
